mysite/requirements.txt
Django==2.1.5
django-taggit==1.1.0
Pillow==7.2.0
pytz==2020.1
whitenoise==5.2.0
